+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/modules/location/Odometry.java
@@ -4,7 +4,6 @@ import androidx.annotation.NonNull;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.teamcode.Movement;
 import org.firstinspires.ftc.teamcode.modules.FieldCentricDriveTrain;
 
 public class Odometry extends FieldCentricDriveTrain implements Locator {
