+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/modules/location/OdometryLocalizer.java
@@ -4,7 +4,6 @@ import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.localization.Localizer;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.teamcode.Movement;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
