+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/autonomous/red/wall/AutonomousRedWallStartMiddleParkImpl.java
