+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/SwitchStatement.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.util;
 
 import android.util.Pair;
